Struct doapi::request::RequestBuilder

pub struct RequestBuilder<'t, T> {
    pub auth: &'t str,
    pub method: Method,
    pub url: String,
    pub resp_t: PhantomData<*const T>,
    pub body: Option<String>,
}

Fields

auth

method

url

resp_t

body

Methods

impl<'t> RequestBuilder<'t, Domains>

fn create(self, name: &str, ip: &str) -> RequestBuilder<'t, Domain>

Returns a request that can be used to create a new domain.

Example

// ... domgr set up same as before
match domgr.domains()
           .create("super.com", "10.10.10.1")
           .retrieve() {
    Ok(domain) => println!("Domain: {}", domain),
    Err(e)     => println!("Error: {}", e)
}

impl<'t> RequestBuilder<'t, Domain>

fn delete(self) -> RequestBuilder<'t, HeaderOnly>

Returns a request that can be used to an existing domain. Returns a header with "status: 204 No Content" if successful

Example

// ... domgr set up same as before
match domgr.domain("super.com")
           .delete()
           .retrieve() {
    Ok(_)  => println!("Success"),
    Err(_) => println!("Error")
}

fn dns_records(self) -> RequestBuilder<'t, DnsRecords>

Returns a request that can be used to list all domains, or perform domain tasks

Example

// ... domgr set up same as before
match domgr.domains()
           .create("super.com", "10.10.10.1")
           .retrieve() {
    Ok(domain) => println!("Domain: {}", domain),
    Err(e)     => println!("Error: {}", e)
}

fn dns_record(self, id: &str) -> RequestBuilder<'t, DnsRecord>

Returns a request that can be used to list a single domain, or perform tasks on a single domain

Example

// ... domgr set up same as before
match domgr.domain("super.com")
           .retrieve() {
    Ok(domain) => println!("Domain: {}", domain),
    Err(e)     => println!("Error: {}", e)
}

impl<'t> RequestBuilder<'t, Droplet>

fn disable_backups(self) -> RequestBuilder<'t, Action>

fn reboot(self) -> RequestBuilder<'t, Action>

fn power_cycle(self) -> RequestBuilder<'t, Action>

fn shutdown(self) -> RequestBuilder<'t, Action>

fn power_off(self) -> RequestBuilder<'t, Action>

fn power_on(self) -> RequestBuilder<'t, Action>

fn restore(self, img: &str) -> RequestBuilder<'t, Action>

fn reset_password(self) -> RequestBuilder<'t, Action>

fn resize(self, size: &str, disk: bool) -> RequestBuilder<'t, Action>

fn rebuild(self, img: &str) -> RequestBuilder<'t, Action>

fn rename(self, name: &str) -> RequestBuilder<'t, Action>

fn change_kernel(self, kernel: &str) -> RequestBuilder<'t, Action>

fn enable_ipv6(self) -> RequestBuilder<'t, Action>

fn enable_private_networking(self) -> RequestBuilder<'t, Action>

fn snapshot(self, name: &str) -> RequestBuilder<'t, Action>

fn upgrade(self) -> RequestBuilder<'t, Action>

fn action(self, id: &str) -> RequestBuilder<'t, Action>

fn kernels(self) -> RequestBuilder<'t, Kernels>

fn snapshots(self) -> RequestBuilder<'t, Snapshots>

fn backups(self) -> RequestBuilder<'t, Backups>

fn actions(self) -> RequestBuilder<'t, Actions>

fn delete(self) -> RequestBuilder<'t, HeaderOnly>

fn show(&self) -> RequestBuilder<'t, Droplet>

fn neighbors(self) -> RequestBuilder<'t, Droplets>

impl<'t> RequestBuilder<'t, Droplets>

fn create(self, droplet: &Droplet) -> RequestBuilder<'t, Droplet>

fn neighbors(self) -> RequestBuilder<'t, Neighbors>

fn upgrades(self) -> RequestBuilder<'t, DropletUpgrades>

impl<'t> RequestBuilder<'t, Image>

fn transfer(self, region: &str) -> RequestBuilder<'t, Action>

fn convert(self) -> RequestBuilder<'t, Action>

fn actions(self) -> RequestBuilder<'t, Actions>

fn action(self, id: &str) -> RequestBuilder<'t, Action>

fn update(self, name: &str) -> RequestBuilder<'t, Image>

fn delete(self) -> RequestBuilder<'t, HeaderOnly>

impl<'t> RequestBuilder<'t, Images>

fn applications(self) -> RequestBuilder<'t, Images>

fn distributions(self) -> RequestBuilder<'t, Images>

fn private(self) -> RequestBuilder<'t, Images>

fn available(self) -> RequestBuilder<'t, Images>

impl<'t> RequestBuilder<'t, SshKeys>

fn create(self, name: &str, pub_key: &str) -> RequestBuilder<'t, SshKey>

impl<'t> RequestBuilder<'t, SshKey>

fn update(self, name: &str) -> RequestBuilder<'t, SshKey>

fn destroy(self) -> RequestBuilder<'t, HeaderOnly>

impl<'t, T> RequestBuilder<'t, T>

fn with_auth(auth: &'t str) -> RequestBuilder<'t, T>

fn new<S>(auth: &'t str, url: S) -> RequestBuilder<'t, T> where S: Into<String>